Add styles to the table
